metrics-util: ensure we always return at least one empty block
For our purposes of decompressing directly into recorders, we always
want to be able to call the underlying method for recording a histogram
at least once, but previously, if we had no values at all,
`decompress_with` would never fire the provided closure, so the metric
would effectively disappear until it had values added to it.

Now, we always return at least one empty block which lets us recognize
the metric and keep it around without meaningfully populating it, not
even with a zero value.